// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface TeamMember {
  id: number;
  name: string;
  position: string;
  image: number | Media;
  socialLinks?:
    | {
        platform: 'instagram' | 'facebook' | 'dribbble' | 'linkedin' | 'twitter';
        url: string;
        id?: string | null;
      }[]
    | null;
  /**
   * Determines display order (ascending)
   */
  order?: number | null;
  /**
   * Used for the team member detail page URL
   */
  slug: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "blog-posts".
 */
export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  featuredImage: number | Media;
  excerpt: string;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  publishedDate: string;
  author?: (number | null) | TeamMember;
  categories?: (number | BlogCategory)[] | null;
  tags?:
    | {
        tag?: string | null;
        /**
         * Optional link for this tag
         */
        link?: string | null;
        id?: string | null;
      }[]
    | null;
  socialLinks?:
    | {
        platform: 'facebook' | 'linkedin' | 'instagram' | 'twitter';
        url: string;
        id?: string | null;
      }[]
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    keywords?: string | null;
    /**
     * Optional canonical URL if this content appears elsewhere
     */
    canonicalURL?: string | null;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "blog-categories".
 */
export interface BlogCategory {
  id: number;
  name: string;
  slug: string;
  updatedAt: string;
  createdAt: string;
}